refactor(backend): move verifier glue out of the spine
Symmetric to the prover-glue move: the Verify trait and its
impl for the bn254 Verifier (which carries a noirc_abi Abi and calls
the bn254 register_ntt) are frontend orchestration, not the generic
verification algorithm. Relocate them to provekit-backend-bn254.

provekit-verifier is reduced to its generic WhirR1CSVerifier<P> spine;
its lib.rs is now just the module + re-export, and it drops the
backend-bn254 dependency. The edge reverses: backend depends on verifier
for the generic verification trait.

Consumers repoint Verify to the backend re-export and drop the now-unused
provekit-verifier dependency. With both glue moves done, cargo tree
confirms prover and verifier depend only on common; backend is the sole
crate that names bn254.

bn254 output is byte-identical (72/72 e2e green).
