+++ provekit/verifier/Cargo.toml
@@ -11,7 +11,6 @@ repository.workspace = true
 [dependencies]
 # Workspace crates
 provekit-common.workspace = true
-provekit-backend-bn254.workspace = true
 
 # Cryptography and proof systems
 ark-ff.workspace = true
