+++ tooling/cli/Cargo.toml
@@ -14,7 +14,6 @@ provekit-common.workspace = true
 provekit-backend-bn254 = { workspace = true, features = ["witness-generation", "parallel"] }
 provekit-gnark.workspace = true
 provekit-r1cs-compiler.workspace = true
-provekit-verifier.workspace = true
 
 # Noir language
 acir.workspace = true
