+++ tooling/provekit-bench/tests/compiler.rs
@@ -4,10 +4,9 @@ use {
     nargo_cli::cli::compile_cmd::compile_workspace_full,
     nargo_toml::{resolve_workspace_from_toml, PackageSelection},
     noirc_driver::CompileOptions,
-    provekit_backend_bn254::{Prove, Prover, Verifier},
+    provekit_backend_bn254::{Prove, Prover, Verifier, Verify},
     provekit_common::HashConfig,
     provekit_r1cs_compiler::NoirCompiler,
-    provekit_verifier::Verify,
     serde::Deserialize,
     std::path::{Path, PathBuf},
     test_case::test_case,
