+++ tooling/provekit-ffi/Cargo.toml
@@ -16,7 +16,6 @@ crate-type = ["staticlib", "rlib"]
 provekit-common.workspace = true
 provekit-backend-bn254 = { workspace = true, features = ["witness-generation", "parallel"] }
 provekit-r1cs-compiler = { workspace = true }
-provekit-verifier = { workspace = true }
 
 # 3rd party
 anyhow.workspace = true
