+++ tooling/provekit-wasm/src/verifier.rs
@@ -1,8 +1,7 @@
 use {
     crate::format::parse_binary_verifier,
-    provekit_backend_bn254::{NoirProof, Verifier as VerifierCore},
+    provekit_backend_bn254::{NoirProof, Verifier as VerifierCore, Verify},
     provekit_common::binary_format::{HEADER_SIZE, MAGIC_BYTES},
-    provekit_verifier::Verify,
     wasm_bindgen::prelude::*,
 };
